Refactored code adding Windows cert store logic, including renaming IISU to WinIIS.
